Maleksm 0.9.434m (04.04.25)

+++ yoRadio/src/core/netserver.h
@@ -78,7 +78,7 @@ class NetServer {
     static void beginUpload(AsyncWebServerRequest *request);
     static void beginUpdate(AsyncWebServerRequest *request);
     void processQueue();
-    int _readPlaylistLine(File &file, char * line, size_t size);
+    uint8_t _readPlaylistLine(File &file, char * line, size_t size);
 };
 
 extern NetServer netserver;
